@@ -1,5 +1,10 @@
-const proxy = require('http-proxy-middleware');
-
-module.exports = function(app) {
-  app.use(proxy('/api', { target: 'http://localhost:5000/' }));
+const { createProxyMiddleware } = require("http-proxy-middleware");
+module.exports = function (app) {
+  app.use(
+    "/api",
+    createProxyMiddleware({
+      target: "http://localhost:5000/",
+      changeOrigin: true,
+    })
+  );
 };
